coworker list status updating

+++ app/src/main/java/com/example/go4lunch/goforlunch/ui/restaurantDetail/RestaurantDetailViewModel.java
@@ -1,6 +1,9 @@
 package com.example.go4lunch.goforlunch.ui.restaurantDetail;
+
 import android.util.Log;
+
 import androidx.lifecycle.MutableLiveData;
+
 import com.example.go4lunch.goforlunch.base.BaseViewModel;
 import com.example.go4lunch.goforlunch.models.Coworker;
 import com.example.go4lunch.goforlunch.models.Restaurant;
@@ -10,8 +13,10 @@ import com.example.go4lunch.goforlunch.utils.Actions;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentSnapshot;
+
 import java.util.ArrayList;
 import java.util.List;
+
 import static com.example.go4lunch.goforlunch.utils.Actions.REMOVED_LIKED;
 import static com.example.go4lunch.goforlunch.utils.Actions.REMOVED_PICKED;
 import static com.example.go4lunch.goforlunch.utils.Actions.UPDATE_LIKED;
@@ -19,8 +24,8 @@ import static com.example.go4lunch.goforlunch.utils.Actions.UPDATE_PICKED;
 
 public class RestaurantDetailViewModel extends BaseViewModel {
 
-    private final CoworkerRepository coworkerRepository ;
-    private final RestaurantRepository restaurantRepository ;
+    private final CoworkerRepository coworkerRepository;
+    private final RestaurantRepository restaurantRepository;
     private Restaurant restaurant;
     private Coworker coworker;
 
@@ -40,11 +45,11 @@ public class RestaurantDetailViewModel extends BaseViewModel {
         return restaurantRepository.getGoogleRestaurantDetail(placeId);
     }
 
-    public void fetchInfoRestaurant(){
+    public void fetchInfoRestaurant() {
         isLoading.setValue(true);
         String uidSelection = restaurantRepository.getRestaurantSelected();
         boolean isRestaurantStored = false;
-        if(restaurantRepository.getRestaurantsLoaded() != null) {
+        if (restaurantRepository.getRestaurantsLoaded() != null) {
             for (Restaurant restaurant : restaurantRepository.getRestaurantsLoaded()) {
                 if (restaurant.getRestaurantPlaceId().equals(uidSelection)) {
                     this.restaurant = restaurant;
@@ -53,17 +58,17 @@ public class RestaurantDetailViewModel extends BaseViewModel {
                 }
             }
         }
-        if(isRestaurantStored){
+        if (isRestaurantStored) {
             fetchCoworkerChoice(restaurant);
 
-        } else{
+        } else {
             getRestaurantDetail(uidSelection);
         }
     }
 
     public void updateRestaurantLiked(Restaurant restaurant) {
         isLoading.setValue(true);
-        if(isRestaurantLiked.getValue()){
+        if (isRestaurantLiked != null && isRestaurantLiked.getValue() != null && isRestaurantLiked.getValue()) {
             coworkerRepository.removeLikedRestaurant(restaurant.getRestaurantPlaceId(), coworker.getUid())
                     .addOnSuccessListener(onSuccessListener(REMOVED_LIKED, restaurant))
                     .addOnFailureListener(this.onFailureListener(UPDATE_LIKED));
@@ -74,26 +79,24 @@ public class RestaurantDetailViewModel extends BaseViewModel {
         }
     }
 
-    public void updatePickedRestaurant(Restaurant res) {
+    public void updatePickedRestaurant(Restaurant restaurant) {
         coworkerRepository.getCoworker(FirebaseAuth.getInstance().getCurrentUser().getUid()).addOnCompleteListener(doc -> {
             coworker = doc.getResult().toObject(Coworker.class);
-            if(isRestaurantPicked != null && isRestaurantPicked.getValue() != null && isRestaurantPicked.getValue()){
+            if (isRestaurantPicked != null && isRestaurantPicked.getValue() != null && isRestaurantPicked.getValue()) {
                 coworkerRepository.updateRestaurantPicked(null, null, null, coworker)
-                        .addOnSuccessListener(onSuccessListener(REMOVED_PICKED, res));
-                        //.addOnFailureListener(this.onFailureListener(UPDATE_PICKED));
+                        .addOnSuccessListener(onSuccessListener(REMOVED_PICKED, restaurant));
             } else {
                 isRestaurantPicked.setValue(true);
-                coworkerRepository.updateRestaurantPicked(res.getRestaurantPlaceId(), res.getRestaurantName(),
-                        res.getRestaurantAddress(), coworker)
-                        .addOnSuccessListener(onSuccessListener(UPDATE_PICKED, res));
-                        //.addOnFailureListener(this.onFailureListener(UPDATE_PICKED));
+                coworkerRepository.updateRestaurantPicked(restaurant.getRestaurantPlaceId(), restaurant.getRestaurantName(),
+                        restaurant.getRestaurantAddress(), coworker)
+                        .addOnSuccessListener(onSuccessListener(UPDATE_PICKED, restaurant));
             }
         });
     }
 
     @Override
     public void action(Actions actions) {
-        switch (actions){
+        switch (actions) {
             case UPDATE_PICKED_RESTAURANT:
                 updatePickedRestaurant(restaurant);
                 break;
@@ -109,17 +112,16 @@ public class RestaurantDetailViewModel extends BaseViewModel {
 
     public void fetchCoworkerChoice(Restaurant restaurant) {
         List<Coworker> userToAdd = new ArrayList<>();
-        //On init le mcoworkerlist en faisant un erase
         ArrayList<Restaurant.CoworkerList> coworkerList = new ArrayList<Restaurant.CoworkerList>();
         mCoworkerList.setValue(coworkerList);
 
         CoworkerRepository.getAllCoworker()
                 .addOnSuccessListener(queryDocumentSnapshots -> {
-                    for (DocumentSnapshot documentSnapshot : queryDocumentSnapshots.getDocuments()){
+                    for (DocumentSnapshot documentSnapshot : queryDocumentSnapshots.getDocuments()) {
                         Coworker coworker = documentSnapshot.toObject(Coworker.class);
-                        if(coworker != null && coworker.getUid() != null){
+                        if (coworker != null && coworker.getUid() != null) {
                             String restaurantUid = coworker.getRestaurantUid();
-                            if(restaurantUid!= null && restaurantUid.equals(restaurant.getRestaurantPlaceId())){
+                            if (restaurantUid != null && restaurantUid.equals(restaurant.getRestaurantPlaceId())) {
                                 userToAdd.add(coworker);
                                 coworkerList.add(new Restaurant.CoworkerList(coworker.getUid(), coworker.getCoworkerName(), coworker.getCoworkerPhotoUrl()));
                             }
@@ -132,26 +134,27 @@ public class RestaurantDetailViewModel extends BaseViewModel {
                 });
     }
 
-    private void configureInfoRestaurant(Restaurant restaurant){
-       // isRestaurantLiked.setValue(checkIfRestaurantIsLiked(restaurant));
+    private void configureInfoRestaurant(Restaurant restaurant) {
+        // isRestaurantLiked.setValue(checkIfRestaurantIsLiked(restaurant));
 
         coworker = coworkerRepository.getActualUser();
         coworkerRepository.getCoworker(FirebaseAuth.getInstance().getCurrentUser().getUid()).addOnCompleteListener(doc -> {
-                    coworker = doc.getResult().toObject(Coworker.class);
-                    String coworkerResUid = coworker.getRestaurantUid();
-                    String restaurantUid = restaurant.getRestaurantPlaceId();
-                    if (coworkerResUid!= null && restaurantUid != null && coworkerResUid.equals(restaurantUid)) {
-                        isRestaurantPicked.setValue(true);
-                    }
-                });
+            coworker = doc.getResult().toObject(Coworker.class);
+            String coworkerResUid = coworker.getRestaurantUid();
+            String restaurantUid = restaurant.getRestaurantPlaceId();
+            isRestaurantLiked.setValue(checkIfRestaurantIsLiked(restaurant));
+            if (coworkerResUid != null && restaurantUid != null && coworkerResUid.equals(restaurantUid)) {
+                isRestaurantPicked.setValue(true);
+            }
+        });
 
 
         isLoading.setValue(false);
     }
 
-    public Boolean checkIfRestaurantIsLiked(Restaurant restaurant){
+    public Boolean checkIfRestaurantIsLiked(Restaurant restaurant) {
         List<String> restaurantLiked = coworker.getCoworkerLikes();
-        if(restaurantLiked != null && restaurantLiked.size() > 0) {
+        if (restaurantLiked != null && restaurantLiked.size() > 0) {
             for (String uid : restaurantLiked) {
                 if (uid.equals(restaurant.getRestaurantPlaceId())) return true;
             }
@@ -159,33 +162,13 @@ public class RestaurantDetailViewModel extends BaseViewModel {
         return false;
     }
 
-    public MutableLiveData<Boolean> checkIfRestaurantIsChosen(Restaurant restaurant){
-//        coworker = coworkerRepository.getActualUser();
-//        Coworker.CoworkerRestaurantChoice restaurantChosen = coworker.getCoworkerRestaurantChosen();
-//        if(restaurantChosen != null) {
-//            if (restaurantChosen.getRestaurantId().equals(restaurant.getRestaurantPlaceId())) return true;
-//        }
-//        return false;
-//        coworkerRepository.getCoworker(FirebaseAuth.getInstance().getCurrentUser().getUid()).addOnCompleteListener(doc -> {
-//            coworker = doc.getResult().toObject(Coworker.class);
-//            if(isRestaurantPicked != null && isRestaurantPicked.getValue() != null && isRestaurantPicked.getValue()){
-//                coworkerRepository.updateRestaurantPicked(null, null, null, coworker)
-//                        .addOnSuccessListener(onSuccessListener(REMOVED_PICKED, restaurant))
-//                        .addOnFailureListener(this.onFailureListener(UPDATE_PICKED));
-//            } else {
-//                isRestaurantPicked.setValue(true);
-//                coworkerRepository.updateRestaurantPicked(restaurant.getRestaurantPlaceId(), restaurant.getRestaurantName(),
-//                        restaurant.getRestaurantAddress(), coworker)
-//                        .addOnSuccessListener(onSuccessListener(UPDATE_PICKED, restaurant))
-//                        .addOnFailureListener(this.onFailureListener(UPDATE_PICKED));
-//            }
-       // });
-return isRestaurantPicked;
+    public MutableLiveData<Boolean> checkIfRestaurantIsChosen(Restaurant restaurant) {
+        return isRestaurantPicked;
     }
 
-    private OnSuccessListener<Void> onSuccessListener(final Actions actions, Restaurant restaurant){
+    private OnSuccessListener<Void> onSuccessListener(final Actions actions, Restaurant restaurant) {
         return aVoid -> {
-            switch (actions){
+            switch (actions) {
                 case UPDATE_PICKED:
                     Log.d("we step in", "onSuccessListener: ");
                     isRestaurantPicked.setValue(true);
